DAOS-16672 control: Move storage_query pretty printers into own file
Required-githooks: true
